Restructure examples order.
