+++ src/InternalAlgorithms/ReflectionAnalysisPoC/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
+using Temporal.Worker.Workflows;
 
 namespace ReflectionAnalysisPoC
 {
