+++ src/InternalAlgorithms/ReflectionAnalysisPoC/ReflectionAnalysisPoC.csproj
@@ -5,4 +5,8 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\Host-SDK\Temporal.Sdk.Worker.csproj" />
+  </ItemGroup>
+
 </Project>
