--- src/Samples/Basic/Temporal.Samples.DesignProcess/Part04_TypicalWorkflow.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Temporal.Sdk.BasicSamples
-{
-    class Part04_TypicalWorkflow
-    {
-    }
-}
